feat(sidebar): reorderRoot handles a member source
When the policy emits .reorderRoot for a dragged account, the source
may now be a current group member (once the policy guard relaxes in
a follow-up commit). Generalise the dispatch to remove the source
from its group first via AccountGroupStore.removeAccount — the
existing helper auto-deletes the old group when emptied.

Also guard the walk-order rewrite against phantom empty-group entries:
when an account observation fires before the matching group-deletion
observation, the snapshot may still include the now-empty group as an
entry. moveGroup on a deleted group throws 404, so skip groups with
no members (a transient race state — groups always have ≥1 member in
normal operation).

Split SidebarDropDispatchReorderTests into ReorderRootTests +
ReorderMembersTests to keep each struct within the 250-line body limit.
